+++ src/Api/Api.csproj
@@ -20,11 +20,7 @@
     </PackageReference>
     <PackageReference Include="Microsoft.Extensions.Diagnostics.HealthChecks.EntityFrameworkCore" Version="8.0.1" />
     <PackageReference Include="MinimalApis.Extensions" Version="0.11.0" />
-      <PackageReference Include="Serilog.AspNetCore" Version="8.0.1" />
-      <PackageReference Include="Serilog.Enrichers.Environment" Version="2.3.0" />
-      <PackageReference Include="Serilog.Expressions" Version="4.0.0" />
-      <PackageReference Include="Serilog.Sinks.Datadog.Logs" Version="0.5.2" />
-      <PackageReference Include="UAParser" Version="3.1.47" />
+    <PackageReference Include="UAParser" Version="3.1.47" />
   </ItemGroup>
 
   <ItemGroup>
